Refactor HealthCheckController methods - move CheckHealth method into CheckHealthService class - move VersionInfo method into CheckHealthService class - declare the above methods in the ICheckHealthService interface - register the CheckHealthService as a transient service
