+++ InvestmentManager/InvestmentManager/Services/ICheckHealthService.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Mvc;
+using System.Threading.Tasks;
+
+namespace InvestmentManager.Services
+{
+    public interface ICheckHealthService
+    {
+        public Task<IActionResult> CheckHealth();
+        public IActionResult VersionInfo();
+
+    }
+}
